Atrac: Rework the header dependencies
