+++ Core/HLE/AtracCtx.cpp
@@ -28,8 +28,7 @@
 #include "Core/HLE/AtracCtx.h"
 #include "Core/HW/Atrac3Standalone.h"
 #include "Core/HLE/sceKernelMemory.h"
-#include <sstream>
-#include <iomanip>
+
 
 const size_t overAllocBytes = 16384;
