+++ Core/HLE/AtracCtx.h
@@ -23,136 +23,8 @@
 #include "Common/Swap.h"
 
 #include "Core/MemMap.h"
-#include "Core/HLE/sceAtrac.h"
-
-struct AtracSingleResetBufferInfo {
-	u32_le writePosPtr;
-	u32_le writableBytes;
-	u32_le minWriteBytes;
-	u32_le filePos;
-};
-
-struct AtracResetBufferInfo {
-	AtracSingleResetBufferInfo first;
-	AtracSingleResetBufferInfo second;
-};
-
-struct AtracSasStreamState {
-	u32 bufPtr[2]{};
-	u32 bufSize[2]{};
-	int streamOffset = 0;
-	int fileOffset = 0;
-	int curBuffer = 0;
-	bool isStreaming = false;
-
-	int CurPos() const {
-		int retval = fileOffset - bufSize[curBuffer] + streamOffset;
-		_dbg_assert_(retval >= 0);
-		return retval;
-	}
-};
-
-const int PSP_ATRAC_ALLDATA_IS_ON_MEMORY = -1;
-const int PSP_ATRAC_NONLOOP_STREAM_DATA_IS_ON_MEMORY = -2;
-const int PSP_ATRAC_LOOP_STREAM_DATA_IS_ON_MEMORY = -3;
-
-// This is not a PSP-native struct.
-// But, it's stored in its entirety in savestates, which makes it awkward to change it.
-// This is used for both first_ and second_, but the latter doesn't use all the fields.
-struct InputBuffer {
-	// Address of the buffer.
-	u32 addr;
-	// Size of data read so far into dataBuf_ (to be removed.)
-	u32 size;
-	// Offset into addr at which new data is added.
-	u32 offset;
-	// Last writableBytes number (to be removed.)
-	u32 writableBytes;
-	// Unused, always 0.
-	u32 neededBytes;
-	// Total size of the entire file data.
-	u32 _filesize_dontuse;
-	// Offset into the file at which new data is read.
-	u32 fileoffset;
-};
-
-class AudioDecoder;
-
-class AtracBase {
-public:
-	virtual ~AtracBase();
-
-	virtual void DoState(PointerWrap &p) = 0;
-
-	// TODO: Find a way to get rid of this from the base class.
-	virtual void UpdateContextFromPSPMem() = 0;
-
-	virtual int Channels() const = 0;
-
-	int GetOutputChannels() const {
-		return outputChannels_;
-	}
-	void SetOutputChannels(int channels) {
-		// Only used for sceSas audio. To be refactored away in the future.
-		outputChannels_ = channels;
-	}
-
-	virtual u32 GetInternalCodecError() const { return 0; }
-
-	PSPPointer<SceAtracContext> context_{};
-
-	virtual AtracStatus BufferState() const = 0;
-
-	virtual int SetLoopNum(int loopNum) = 0;
-	virtual int LoopNum() const = 0;
-	virtual int LoopStatus() const = 0;
-
-	virtual int CodecType() const = 0;
-
-	AudioDecoder *Decoder() const {
-		return decoder_;
-	}
-
-	void CreateDecoder(int codecType, int bytesPerFrame, int channels);
-
-	virtual void NotifyGetContextAddress() = 0;
-
-	virtual int GetNextDecodePosition(int *pos) const = 0;
-	virtual int RemainingFrames() const = 0;
-	virtual bool HasSecondBuffer() const = 0;
-	virtual int Bitrate() const = 0;
-	virtual int BytesPerFrame() const = 0;
-	virtual int SamplesPerFrame() const = 0;
-
-	virtual void GetStreamDataInfo(u32 *writePtr, u32 *writableBytes, u32 *readOffset) = 0;  // This should be const, but the legacy impl stops it (it's wrong).
-	virtual int AddStreamData(u32 bytesToAdd) = 0;
-	virtual int ResetPlayPosition(int sample, int bytesWrittenFirstBuf, int bytesWrittenSecondBuf, bool *delay) = 0;
-	virtual int GetBufferInfoForResetting(AtracResetBufferInfo *bufferInfo, int sample, bool *delay) = 0;  // NOTE: Not const! This can cause SkipFrames!
-	virtual int SetData(const Track &track, u32 buffer, u32 readSize, u32 bufferSize, int outputChannels) = 0;
-
-	virtual int GetSecondBufferInfo(u32 *fileOffset, u32 *desiredSize) const = 0;
-	virtual int SetSecondBuffer(u32 secondBuffer, u32 secondBufferSize) = 0;
-	virtual u32 DecodeData(u8 *outbuf, u32 outbufPtr, int *SamplesNum, int *finish, int *remains) = 0;
-	virtual int DecodeLowLevel(const u8 *srcData, int *bytesConsumed, s16 *dstData, int *bytesWritten) = 0;
-
-	virtual u32 GetNextSamples() = 0;  // This should be const, but the legacy impl stops it (it's wrong).
-	virtual void InitLowLevel(const Atrac3LowLevelParams &params, int codecType) = 0;
-
-	virtual void CheckForSas() = 0;
-	virtual int EnqueueForSas(u32 address, u32 ptr) = 0;
-	virtual void DecodeForSas(s16 *dstData, int *bytesWritten, int *finish) = 0;
-	virtual const AtracSasStreamState *StreamStateForSas() const { return nullptr; }
-
-	virtual int GetSoundSample(int *endSample, int *loopStartSample, int *loopEndSample) const = 0;
-
-	virtual int GetContextVersion() const = 0;
-
-protected:
-	u16 outputChannels_ = 2;
-
-	// TODO: Save the internal state of this, now technically possible.
-	AudioDecoder *decoder_ = nullptr;
-};
+#include "Core/HLE/AtracBase.h"
+#include "Core/Util/AtracTrack.h"
 
 class Atrac : public AtracBase {
 public:
